+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Change
+
+- Update route53 trust identity policy with IRSA to account for `cert-manager-controller`
+
 ## [0.6.0] - 2023-02-17
 
 ### Added